requirements.txt
python-telegram-bot
Django=="3.2.9"
python-dotenv
requests
datetime
psycopg2
gunicorn
pipenv